added routing, sev. components, cancel-button, players' count
