+++ client/src/pages/Animals.jsx
@@ -1,14 +1,8 @@
-// import { NavLink, Routes, Route } from 'react-router-dom';
-// import App from "../App"
-
 export default function Animals() {
     return (
       <>
-        {/* <NavLink to="/">Home</NavLink>
-        <Routes>
-          <Route path="/" element={<App/>}/>
-        </Routes> */}
         <h3>Gallery of Animals</h3>
+        <p><em>Animals to follow soon...</em></p>
       </>
     );
   }
