+++ client/src/pages/Disciplines.jsx
@@ -1,15 +1,8 @@
-// import { NavLink, Routes, Route } from 'react-router-dom';
-// import App from "../App"
-
-
 export default function Disciplines() {
     return (
       <>
-        {/* <NavLink to="/">Home</NavLink>
-        <Routes>
-          <Route path="/" element={<App/>}/>
-        </Routes> */}
          <h3>Gallery of Disciplines</h3>
+         <p><em>Disciplines to follow soon...</em></p>
       </>
     );
   }
