+++ client/src/pages/Info.jsx
@@ -1,13 +1,6 @@
-// import { NavLink, Routes, Route } from 'react-router-dom';
-// import App from "../App"
-
 export default function Info() {
     return (
       <>
-        {/* <NavLink to="/">Home</NavLink>
-        <Routes>
-          <Route path="/" element={<App/>}/>
-        </Routes> */}
          <h3>About the Game</h3>
          <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Laudantium omnis totam veritatis sed fugit voluptatem praesentium quas repellat autem expedita consectetur eveniet, obcaecati voluptas illum facilis nisi maxime rerum est perferendis animi a quis vitae? Commodi expedita facere hic ipsam! Magni vero, id rem iste dolore veritatis voluptate voluptatibus repellendus.</p>
       </>
